refactor(cli): split selfupdate into version, download and install
selfupdate.go 537 -> 123, which is now only the package-manager detection and
Apply, the orchestration that reads top to bottom. The three steps it calls
move to files named for them:

- version.go: what is newest, and the on-disk cache that keeps the CLI from
  asking GitHub on every invocation.
- download.go: fetching a release asset and turning it into a binary, from
  asset naming through checksum verification and archive extraction.
- install.go: replacing the running executable, the step that differs most
  between platforms.

install.go is named install.go and not install_windows.go on purpose. A GOOS
or GOARCH suffix is an IMPLICIT build constraint in Go: a file named
install_windows.go would silently vanish from the Linux and macOS builds, and
nothing on a Windows developer machine would notice. The file comment says so,
because the next person to add a platform branch here is the one at risk.

This is the one split in the program where the platform proof is not
theoretical, so it is run three times:

  for os in windows linux darwin; do
    diff <(cd ../floe-base/cli && GOOS=$os go doc -all -u ./internal/selfupdate) \
         <(cd cli && GOOS=$os go doc -all -u ./internal/selfupdate)
  done

All three are empty, all three agree with each other, and all three build. A
stray GOOS suffix would show up as a shrunken FUNCTIONS list on two of the
three.

Test plan
- go doc -all -u diff against the base worktree, per GOOS: empty for windows,
  linux and darwin, and the three outputs are identical to each other.
- GOOS=windows, linux and darwin all build.
- All 20 moved function bodies md5-identical to their base copies.
- go vet and go test clean in cli/.
- go doc cannot see cobra wiring, so the help text of all five commands
  (root, send, receive, version, update) is diffed against a binary built
  from the base worktree. All five identical.
